pyproject.toml
[tool.poetry]
name = "gradescopescheduler"
version = "0.1.0"
description = ""
authors = ["repl.it user <replituser@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
bs4 = "^0.0.1"
google-api-python-client = "^1.12.8"
google-auth-oauthlib = "^0.4.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
